# artifacts/api-server/src/routes/chat.ts
import { Router, type IRouter } from "express";
import { db, jobsTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";
import {
  CreateChatCompletionBody,
  CreateChatCompletionResponse,
  ListChatJobsResponse,
} from "@workspace/api-zod";
import { requireAuth, type AuthedRequest } from "../lib/auth";
import { chargeAndCreateJob, recordJobCompletion } from "../lib/billing";
import { MODEL_CREDITS, MODEL_TIER } from "../lib/constants";
import { openrouter } from "@workspace/integrations-openrouter-ai";
import { openai, generateImageBuffer } from "@workspace/integrations-openai-ai-server";
import { anthropic } from "@workspace/integrations-anthropic-ai";
import { chatRateLimit } from "../lib/security";
import { BEE_SYSTEM_PROMPT, FREE_MODELS, MAX_TOKENS, IMAGE_GEN_CREDITS, isImageRequest } from "../lib/ai";
import { storeImage, getImage } from "../lib/image-cache";
import type { Request, Response } from "express";

const router: IRouter = Router();

// REST endpoint for non-socket clients (curl, third-party integrations, etc.)
router.post(
  "/chat/completions",
  chatRateLimit,
  requireAuth,
  async (req, res): Promise<void> => {
    const user = (req as AuthedRequest).user;
    const parsed = CreateChatCompletionBody.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.message });
      return;
    }

    const { model, messages } = parsed.data;
    const lastUserMsg = [...messages].reverse().find((m) => m.role === "user")?.content ?? "";

    // ── Image generation path ─────────────────────────────────────────────
    if (isImageRequest(lastUserMsg)) {
      const creditsRequired = IMAGE_GEN_CREDITS;
      const tier = "hover";

      const charge = await chargeAndCreateJob({ userId: user.id, tier, model, creditsRequired });
      if (!charge.ok) {
        res.status(402).json({ error: "Insufficient credits. Please top up your balance." });
        return;
      }

      const startedAt = Date.now();
      try {
        const imgBuffer = await generateImageBuffer(lastUserMsg, "1024x1024");
        const imgId = crypto.randomUUID();
        storeImage(imgId, imgBuffer);
        const content = `![Generated Image](/api/chat/image/${imgId})\n\n*Image generated by Bee · ${IMAGE_GEN_CREDITS} credits*`;

        await recordJobCompletion({
          jobId: charge.jobId,
          creditsCharged: creditsRequired,
          promptTokens: Math.ceil(lastUserMsg.length / 4),
          completionTokens: 0,
          ttfbMs: Date.now() - startedAt,
          durationMs: Date.now() - startedAt,
        });

        res.json(
          CreateChatCompletionResponse.parse({
            id: charge.jobId,
            model,
            content,
            creditsDeducted: creditsRequired,
            workerId: "hive-dalle3",
            tokensUsed: Math.ceil(lastUserMsg.length / 4),
          }),
        );
      } catch (err) {
        req.log.error({ err }, "DALL-E 3 image generation failed");
        res.status(502).json({ error: "Image generation failed. Please try again." });
      }
      return;
    }

    const creditsRequired = MODEL_CREDITS[model] ?? 10;
    const tier = MODEL_TIER[model] ?? "hover";

    const charge = await chargeAndCreateJob({
      userId: user.id,
      tier,
      model,
      creditsRequired,
    });

    if (!charge.ok) {
      res.status(402).json({ error: "Insufficient credits. Please top up your balance." });
      return;
    }

    const startedAt = Date.now();
    const promptTokens = messages.reduce(
      (s, m) => s + Math.ceil(m.content.length / 4),
      0,
    );

    let content = "";
    let firstTokenAt: number | undefined;
    let inferenceOk = false;

    // ── 1. Primary: OpenAI integration ───────────────────────────────────
    try {
      const stream = await openai.chat.completions.create({
        model: "gpt-4.1",
        max_tokens: MAX_TOKENS,
        stream: true,
        messages: [
          { role: "system", content: BEE_SYSTEM_PROMPT },
          ...messages.map((m) => ({ role: m.role as "user" | "assistant", content: m.content })),
        ],
      });

      for await (const chunk of stream) {
        const delta = chunk.choices[0]?.delta?.content;
        if (delta) {
          if (!firstTokenAt) firstTokenAt = Date.now();
          content += delta;
        }
      }
      inferenceOk = true;
    } catch (err) {
      req.log.warn({ err }, "OpenAI integration failed — falling back to free models");
    }

    // ── 2. Fallback: Claude Sonnet (Anthropic) ────────────────────────────
    if (!inferenceOk) {
      try {
        const claudeMessages = messages.map((m) => ({
          role: m.role as "user" | "assistant",
          content: m.content,
        }));
        const stream = anthropic.messages.stream({
          model: "claude-sonnet-4-6",
          max_tokens: MAX_TOKENS,
          system: BEE_SYSTEM_PROMPT,
          messages: claudeMessages,
        });
        for await (const event of stream) {
          if (
            event.type === "content_block_delta" &&
            event.delta.type === "text_delta"
          ) {
            if (!firstTokenAt) firstTokenAt = Date.now();
            content += event.delta.text;
          }
        }
        inferenceOk = true;
      } catch (err) {
        req.log.warn({ err }, "Claude Sonnet fallback failed — trying OpenRouter");
      }
    }

    // ── 3. Fallback: OpenRouter free models ──────────────────────────────
    if (!inferenceOk) {
      for (const candidate of FREE_MODELS) {
        try {
          const stream = await openrouter.chat.completions.create({
            model: candidate,
            max_tokens: MAX_TOKENS,
            stream: true,
            messages: [
              { role: "system", content: BEE_SYSTEM_PROMPT },
              ...messages.map((m) => ({ role: m.role as "user" | "assistant", content: m.content })),
            ],
          });

          for await (const chunk of stream) {
            const delta = chunk.choices[0]?.delta?.content;
            if (delta) {
              if (!firstTokenAt) firstTokenAt = Date.now();
              content += delta;
            }
          }
          inferenceOk = true;
          break;
        } catch (err) {
          const status = (err as { status?: number }).status;
          if (status === 404) {
            req.log.warn({ candidate }, "Model unavailable (404), skipping");
          } else if (status === 429) {
            req.log.warn({ candidate }, "Model rate-limited (429), trying next");
          } else {
            req.log.warn({ err, candidate }, "Model error, trying next");
          }
        }
      }
    }

    if (!inferenceOk) {
      res.status(502).json({ error: "Inference failed. Please try again." });
      return;
    }

    const completionTokens = Math.ceil(content.length / 4);
    const durationMs = Date.now() - startedAt;
    const ttfbMs = firstTokenAt ? firstTokenAt - startedAt : durationMs;

    await recordJobCompletion({
      jobId: charge.jobId,
      creditsCharged: creditsRequired,
      promptTokens,
      completionTokens,
      ttfbMs,
      durationMs,
    });

    res.json(
      CreateChatCompletionResponse.parse({
        id: charge.jobId,
        model,
        content,
        creditsDeducted: creditsRequired,
        workerId: "hive-" + Math.random().toString(36).slice(2, 10),
        tokensUsed: promptTokens + completionTokens,
      }),
    );
  },
);

router.get("/chat/jobs", requireAuth, async (req, res): Promise<void> => {
  const user = (req as AuthedRequest).user;
  const jobs = await db
    .select()
    .from(jobsTable)
    .where(eq(jobsTable.userId, user.id))
    .orderBy(desc(jobsTable.createdAt))
    .limit(50);

  const terminal = ["done", "failed", "refunded"];
  res.json(
    ListChatJobsResponse.parse(
      jobs
        .filter((j) => terminal.includes(j.status))
        .map((j) => ({
          id: j.id,
          model: j.model,
          tier: j.tier,
          creditsSpent: j.creditsCharged,
          tokensUsed: (j.promptTokens ?? 0) + (j.completionTokens ?? 0),
          status: j.status === "done" ? "completed" : j.status,
          workerId: j.workerId ?? null,
          createdAt: j.createdAt.toISOString(),
        })),
    ),
  );
});

// Serve generated images from in-memory cache (no auth — URLs are unguessable UUIDs)
router.get("/chat/image/:id", (req: Request, res: Response): void => {
  const buf = getImage(String(req.params.id ?? ""));
  if (!buf) {
    res.status(404).json({ error: "Image not found or expired." });
    return;
  }
  res.setHeader("Content-Type", "image/png");
  res.setHeader("Cache-Control", "public, max-age=900");
  res.end(buf);
});

export default router;