# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gigabee decentralized AI inference network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/magic-link:
    post:
      operationId: requestMagicLink
      tags: [auth]
      summary: Request a magic link email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicLinkResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    post:
      operationId: verifyMagicLink
      tags: [auth]
      summary: Verify magic link token and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkVerify"
      responses:
        "200":
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/completions:
    post:
      operationId: createChatCompletion
      tags: [chat]
      summary: Create a chat completion (OpenAI-compatible)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/jobs:
    get:
      operationId: listChatJobs
      tags: [chat]
      summary: Get recent chat job metadata (no content, only metadata)
      responses:
        "200":
          description: Recent chat jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatJob"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api-keys:
    get:
      operationId: listApiKeys
      tags: [auth]
      summary: List active API keys
      description: Returns all non-revoked API keys for the authenticated user (key prefix only, never the raw key).
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createApiKey
      tags: [auth]
      summary: Create a new API key
      description: Generates a new API key. The raw key is returned once and cannot be retrieved again.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
        "400":
          description: Key limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [auth]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Key not found or already revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    MagicLinkResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    MagicLinkVerify:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AuthSession:
      type: object
      required: [user, sessionToken]
      properties:
        user:
          $ref: "#/components/schemas/User"
        sessionToken:
          type: string
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          enum: [bee-nano, bee-hover, bee-glide]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
    ChatCompletionResponse:
      type: object
      required: [id, model, content, creditsDeducted, workerId, tokensUsed]
      properties:
        id:
          type: string
        model:
          type: string
        content:
          type: string
        creditsDeducted:
          type: integer
        workerId:
          type: string
        tokensUsed:
          type: integer
    ChatJob:
      type: object
      required: [id, model, tier, creditsSpent, tokensUsed, status, createdAt]
      properties:
        id:
          type: string
        model:
          type: string
        tier:
          type: string
        creditsSpent:
          type: integer
        tokensUsed:
          type: integer
        status:
          type: string
          enum: [completed, failed, refunded]
        workerId:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ApiKey:
      type: object
      required: [id, keyPrefix, createdAt]
      properties:
        id:
          type: string
        keyPrefix:
          type: string
        label:
          type: ["string", "null"]
        createdAt:
          type: string
        lastUsedAt:
          type: ["string", "null"]
    ApiKeyList:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
    CreateApiKeyRequest:
      type: object
      properties:
        label:
          type: string
    ApiKeyCreated:
      type: object
      required: [key, id, keyPrefix, createdAt]
      properties:
        key:
          type: string
        id:
          type: string
        keyPrefix:
          type: string
        label:
          type: ["string", "null"]
        createdAt:
          type: string